cloud_functions_src/requirements.txt
google-cloud-resource-manager
google-cloud-tasks
google-cloud-billing
